Restructured examples
