+++ sdkAngular/app/dataform/styling/advanced/dataform-styling-advanced.component.html
