+++ sdkAngular/app/dataform/styling/advanced/dataform-styling-advanced.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { Settings } from "../data-services/settings";
+import { Settings } from "../../data-services/settings";
 import * as applicationModule from "tns-core-modules/application";
 import { Color } from "tns-core-modules/color";
 import * as viewModule from "tns-core-modules/ui/core/view";
