+++ sdkAngular/app/dataform/styling/common/dataform-styling.component.html
